+++ proxy/frontend/postgresql/src/main/java/org/apache/shardingsphere/proxy/frontend/postgresql/command/query/extended/execute/PostgreSQLComExecuteExecutor.java
@@ -25,6 +25,7 @@ import org.apache.shardingsphere.proxy.frontend.postgresql.command.query.extende
 import org.apache.shardingsphere.sql.parser.sql.common.statement.tcl.CommitStatement;
 import org.apache.shardingsphere.sql.parser.sql.common.statement.tcl.RollbackStatement;
 
+import java.sql.SQLException;
 import java.util.List;
 
 /**
@@ -39,13 +40,13 @@ public final class PostgreSQLComExecuteExecutor implements CommandExecutor {
     
     @SuppressWarnings({"rawtypes", "unchecked"})
     @Override
-    public List execute() {
+    public List execute() throws SQLException {
         return portalContext.get(packet.getPortal()).execute(packet.getMaxRows());
     }
     
     @Override
-    public void close() {
-        Portal<?> portal = portalContext.get(packet.getPortal());
+    public void close() throws SQLException {
+        Portal portal = portalContext.get(packet.getPortal());
         if (portal.getSqlStatement() instanceof CommitStatement || portal.getSqlStatement() instanceof RollbackStatement) {
             portalContext.closeAll();
         }
